updated ZFP to 1.0.0

+++ src/lib.rs
@@ -38,9 +38,9 @@ mod zfp {
             zfp_field_3d(
                 array.as_mut_ptr() as *mut std::ffi::c_void,
                 data_type,
-                nx as u32,
-                ny as u32,
-                nz as u32,
+                nx as usize,
+                ny as usize,
+                nz as usize,
             )
         };
